+++ ta_codegen/output/rust/library/src/ta_func/ht_trendmode.rs
@@ -247,7 +247,7 @@ impl Core {
         let mut smoothPrice_Idx: usize = 0;
         let mut maxIdx_smoothPrice: usize = 49;
         let mut idx: usize = 0_usize;
-        let mut DCPeriodInt: usize = 0_usize;
+        let mut DCPeriodInt: i32 = 0_i32;
         let mut DCPhase: f64 = 0.0_f64;
         let mut DCPeriod: f64 = 0.0_f64;
         let mut imagPart: f64 = 0.0_f64;
@@ -547,15 +547,15 @@ impl Core {
             // Compute Dominant Cycle Phase
             prevDCPhase = DCPhase;
             DCPeriod = smoothPeriod + 0.5;
-            DCPeriodInt = (DCPeriod as usize) as usize;
+            DCPeriodInt = (DCPeriod) as i32;
             realPart = 0.0;
             imagPart = 0.0;
             // idx is used to iterate for up to 50 of the last
             // value of smoothPrice.
             idx = smoothPrice_Idx;
-            // for( i = 0; i < DCPeriodInt; i += 1 )
+            // for( i = 0; ((i) as i32) < DCPeriodInt; i += 1 )
             i = 0;
-            while i < DCPeriodInt {
+            while ((i) as i32) < DCPeriodInt {
                 tempReal = (i as f64) * constDeg2RadBy360 / (DCPeriodInt as f64);
                 tempReal2 = smoothPrice[idx];
                 realPart += (tempReal).sin() * tempReal2;
@@ -592,7 +592,7 @@ impl Core {
             leadSine = ((DCPhase + 45_f64) * deg2Rad).sin();
             // Compute Trendline
             DCPeriod = smoothPeriod + 0.5;
-            DCPeriodInt = (DCPeriod as usize) as usize;
+            DCPeriodInt = (DCPeriod) as i32;
             // Average the RAW price over the dominant cycle period.
             // Unlike the DC-phase loop above (which reads the smoothPrice
             // circular buffer), the iTrend average reads the raw price,
@@ -608,7 +608,7 @@ impl Core {
             // for( j = 0; j < 50; j += 1 )
             j = 0;
             while j < 50 {
-                if j < DCPeriodInt {
+                if ((j) as i32) < DCPeriodInt {
                     tempReal += inReal[today - j];
                 }
                 j += 1;
@@ -763,7 +763,7 @@ impl Core {
         let mut smoothPrice_Idx: usize = 0;
         let mut maxIdx_smoothPrice: usize = 49;
         let mut idx: usize = 0_usize;
-        let mut DCPeriodInt: usize = 0_usize;
+        let mut DCPeriodInt: i32 = 0_i32;
         let mut DCPhase: f64 = 0.0_f64;
         let mut DCPeriod: f64 = 0.0_f64;
         let mut imagPart: f64 = 0.0_f64;
@@ -1015,13 +1015,13 @@ impl Core {
             smoothPeriod = (0.67 as f64).mul_add(smoothPeriod, 0.33 * period);
             prevDCPhase = DCPhase;
             DCPeriod = smoothPeriod + 0.5;
-            DCPeriodInt = (DCPeriod as usize) as usize;
+            DCPeriodInt = (DCPeriod) as i32;
             realPart = 0.0;
             imagPart = 0.0;
             idx = smoothPrice_Idx;
-            // for( i = 0; i < DCPeriodInt; i += 1 )
+            // for( i = 0; ((i) as i32) < DCPeriodInt; i += 1 )
             i = 0;
-            while i < DCPeriodInt {
+            while ((i) as i32) < DCPeriodInt {
                 tempReal = (i as f64) * constDeg2RadBy360 / (DCPeriodInt as f64);
                 tempReal2 = smoothPrice[idx];
                 realPart += (tempReal).sin() * tempReal2;
@@ -1056,12 +1056,12 @@ impl Core {
             sine = (DCPhase * deg2Rad).sin();
             leadSine = ((DCPhase + 45_f64) * deg2Rad).sin();
             DCPeriod = smoothPeriod + 0.5;
-            DCPeriodInt = (DCPeriod as usize) as usize;
+            DCPeriodInt = (DCPeriod) as i32;
             tempReal = 0.0;
             // for( j = 0; j < 50; j += 1 )
             j = 0;
             while j < 50 {
-                if j < DCPeriodInt {
+                if ((j) as i32) < DCPeriodInt {
                     tempReal += inReal[today - j];
                 }
                 j += 1;
@@ -1177,7 +1177,7 @@ struct HtTrendModeStreamState {
     constDeg2RadBy360: f64,
     smoothPeriod: f64,
     idx: usize,
-    DCPeriodInt: usize,
+    DCPeriodInt: i32,
     DCPhase: f64,
     DCPeriod: f64,
     imagPart: f64,
@@ -1353,15 +1353,15 @@ impl Core {
         // Compute Dominant Cycle Phase
         sp.prevDCPhase = sp.DCPhase;
         sp.DCPeriod = sp.smoothPeriod + 0.5;
-        sp.DCPeriodInt = (sp.DCPeriod as usize) as usize;
+        sp.DCPeriodInt = (sp.DCPeriod) as i32;
         sp.realPart = 0.0;
         sp.imagPart = 0.0;
         // idx is used to iterate for up to 50 of the last
         // value of smoothPrice.
         sp.idx = sp.smoothPrice_Idx;
-        // for( sp.i = 0; sp.i < sp.DCPeriodInt; sp.i += 1 )
+        // for( sp.i = 0; ((sp.i) as i32) < sp.DCPeriodInt; sp.i += 1 )
         sp.i = 0;
-        while sp.i < sp.DCPeriodInt {
+        while ((sp.i) as i32) < sp.DCPeriodInt {
             sp.tempReal = (sp.i as f64) * sp.constDeg2RadBy360 / (sp.DCPeriodInt as f64);
             sp.tempReal2 = sp.cb_smoothPrice[sp.idx];
             sp.realPart += (sp.tempReal).sin() * sp.tempReal2;
@@ -1398,7 +1398,7 @@ impl Core {
         sp.leadSine = ((sp.DCPhase + 45_f64) * sp.deg2Rad).sin();
         // Compute Trendline
         sp.DCPeriod = sp.smoothPeriod + 0.5;
-        sp.DCPeriodInt = (sp.DCPeriod as usize) as usize;
+        sp.DCPeriodInt = (sp.DCPeriod) as i32;
         // Average the RAW price over the dominant cycle period.
         // Unlike the DC-phase loop above (which reads the smoothPrice
         // circular buffer), the iTrend average reads the raw price,
@@ -1414,7 +1414,7 @@ impl Core {
         // for( sp.j = 0; sp.j < 50; sp.j += 1 )
         sp.j = 0;
         while sp.j < 50 {
-            if sp.j < sp.DCPeriodInt {
+            if ((sp.j) as i32) < sp.DCPeriodInt {
                 sp.tempReal += sp.win_j_inReal[((if sp.winPos_j + sp.winCap_j - sp.j >= sp.winCap_j { sp.winPos_j + sp.winCap_j - sp.j - sp.winCap_j } else { sp.winPos_j + sp.winCap_j - sp.j })) as usize];
             }
             sp.j += 1;
@@ -1547,7 +1547,7 @@ impl Core {
         let mut smoothPrice_Idx: usize = 0;
         let mut maxIdx_smoothPrice: usize = 49;
         let mut idx: usize = 0_usize;
-        let mut DCPeriodInt: usize = 0_usize;
+        let mut DCPeriodInt: i32 = 0_i32;
         let mut DCPhase: f64 = 0.0_f64;
         let mut DCPeriod: f64 = 0.0_f64;
         let mut imagPart: f64 = 0.0_f64;
@@ -1847,15 +1847,15 @@ impl Core {
             // Compute Dominant Cycle Phase
             prevDCPhase = DCPhase;
             DCPeriod = smoothPeriod + 0.5;
-            DCPeriodInt = (DCPeriod as usize) as usize;
+            DCPeriodInt = (DCPeriod) as i32;
             realPart = 0.0;
             imagPart = 0.0;
             // idx is used to iterate for up to 50 of the last
             // value of smoothPrice.
             idx = smoothPrice_Idx;
-            // for( i = 0; i < DCPeriodInt; i += 1 )
+            // for( i = 0; ((i) as i32) < DCPeriodInt; i += 1 )
             i = 0;
-            while i < DCPeriodInt {
+            while ((i) as i32) < DCPeriodInt {
                 tempReal = (i as f64) * constDeg2RadBy360 / (DCPeriodInt as f64);
                 tempReal2 = smoothPrice[idx];
                 realPart += (tempReal).sin() * tempReal2;
@@ -1892,7 +1892,7 @@ impl Core {
             leadSine = ((DCPhase + 45_f64) * deg2Rad).sin();
             // Compute Trendline
             DCPeriod = smoothPeriod + 0.5;
-            DCPeriodInt = (DCPeriod as usize) as usize;
+            DCPeriodInt = (DCPeriod) as i32;
             // Average the RAW price over the dominant cycle period.
             // Unlike the DC-phase loop above (which reads the smoothPrice
             // circular buffer), the iTrend average reads the raw price,
@@ -1908,7 +1908,7 @@ impl Core {
             // for( j = 0; j < 50; j += 1 )
             j = 0;
             while j < 50 {
-                if j < DCPeriodInt {
+                if ((j) as i32) < DCPeriodInt {
                     tempReal += inReal[today - j];
                 }
                 j += 1;
@@ -2165,7 +2165,7 @@ impl Core {
         let mut smoothPrice_Idx: usize = 0;
         let mut maxIdx_smoothPrice: usize = 49;
         let mut idx: usize = 0_usize;
-        let mut DCPeriodInt: usize = 0_usize;
+        let mut DCPeriodInt: i32 = 0_i32;
         let mut DCPhase: f64 = 0.0_f64;
         let mut DCPeriod: f64 = 0.0_f64;
         let mut imagPart: f64 = 0.0_f64;
@@ -2465,15 +2465,15 @@ impl Core {
             // Compute Dominant Cycle Phase
             prevDCPhase = DCPhase;
             DCPeriod = smoothPeriod + 0.5;
-            DCPeriodInt = (DCPeriod as usize) as usize;
+            DCPeriodInt = (DCPeriod) as i32;
             realPart = 0.0;
             imagPart = 0.0;
             // idx is used to iterate for up to 50 of the last
             // value of smoothPrice.
             idx = smoothPrice_Idx;
-            // for( i = 0; i < DCPeriodInt; i += 1 )
+            // for( i = 0; ((i) as i32) < DCPeriodInt; i += 1 )
             i = 0;
-            while i < DCPeriodInt {
+            while ((i) as i32) < DCPeriodInt {
                 tempReal = (i as f64) * constDeg2RadBy360 / (DCPeriodInt as f64);
                 tempReal2 = smoothPrice[idx];
                 realPart += (tempReal).sin() * tempReal2;
@@ -2510,7 +2510,7 @@ impl Core {
             leadSine = ((DCPhase + 45_f64) * deg2Rad).sin();
             // Compute Trendline
             DCPeriod = smoothPeriod + 0.5;
-            DCPeriodInt = (DCPeriod as usize) as usize;
+            DCPeriodInt = (DCPeriod) as i32;
             // Average the RAW price over the dominant cycle period.
             // Unlike the DC-phase loop above (which reads the smoothPrice
             // circular buffer), the iTrend average reads the raw price,
@@ -2526,7 +2526,7 @@ impl Core {
             // for( j = 0; j < 50; j += 1 )
             j = 0;
             while j < 50 {
-                if j < DCPeriodInt {
+                if ((j) as i32) < DCPeriodInt {
                     tempReal += inReal[today - j];
                 }
                 j += 1;
